UI: restructure view hierarchy/event handling
Use subclassing to control the window proc for delegate dispatching.
Clean up the delegate handling to allow for easier handling of self.
This can be sunk into view again, but it makes it clearer to see the
structure of the view hierarchy currently.  Remove now unneeded code
in the tree.