UI: send a `SetWindowPos` on reparenting
When a window is reparented, we need to send the child a `SetWindowPos`
(i.e. `WM_WINDOWPOSCHANGED`).  This improves the rendering of the
`TextField`, although gets its height incorrect currently.  The
`SWP_FRAMECHANGED` will trigger a `WM_NCCALCSIZE` which is required to
ensure that the child is sized properly.  The `SWP_NOMOVE` will ignore
the `x` and `y` coordinates.  `SWP_NOSIZE` ignores `cx` and `cy` at the
cost of incorrectly rendering the child's height.