UI: generalize the windows handle management
Create a new `ManagedHandle` type that represents a managed handle.  The
handle is explicitly marked as owning or referencing.  An owning handle
will be deallocated at the end of the lifetime of the object.  The
constructor explicitly indicates whether the type is owned or not.

This management allows for proper handle type specific handling of the
handles that are being trafficked across the C/Swift boundary.