UI: preserve `frame` and scale as appropriate
This adjusts the handling of the `frame` property to preserve the user
specified values.  The frame is scaled at the time of re-parenting and
adjusted for overlapped windows if created with `CW_USEDEFAULT`.  This
removes the unnecessary computation on the construction of the view as
the frame is not adjusted for the DPI adjusted values.