View: adjust the frame and reset when reparenting
Use the newly minted functionality of setting the frame to resize the
window when reparenting.  This allows for the re-use of the logic rather
than duplicating it.  There is still room to improve the remapping of
the corrdinates when re-parenting but this seems to function
sufficiently for now.

Ensure that an overlapped window is never re-parented as that is
ill-defined.  Explicitly remove the `WS_POPUP` style when reparenting.
`SetParent` will not adjust the window style, we must explicitly remove
the `WS_POPUP` attribute (in case the window was previously parented by
the desktop window) and set the `WS_CHILD` attribute.  Reorder the
handling to update the style being invoking `SetWindowPos` to refresh
the frame after the adjusted window properties (`SetWindowLongPtr`).