UI: add support to construct monospaced fonts
Add constructors for retrieving the system default monospaced fonts.  As
Windows does not have the concept of a system default font (unlike Linux
where FontConfig will provide the ability to set the default font for
the font family), we emulate that by specifying NULL for the family
name, which allows the system to provide the first font that matches the
remaining constraints.