UI: enable mapping the window to the screen
This adds the monitor handle to the screen to allow mapping the screen
to the monitor.

Add the helper `==` operators for comparing a screen to a `HMONITOR`.
These are internal as the user should not need to be aware of the
Windows types.

Add the screen property to the Window type.  This allows querying the
screen from the Window.