UI: start wiring up `WM_DPICHANGED` notifications
Process the `WM_DPICHANGED` message in the subclass'ed window procedure.
This allows us to identify the screen (monitor) that the event is
associated with and query the new resolution on the screen to adjust the
`traitCollection` property as a `TraitEnvironment`.

The screen should in the future broadcast this back to all hypothetical
view controllers associated with the windows on the particular screen
which can then broadcast this down the chain of views to re-layout the
elements via the `TraitEnvironment` handling mechanism
(`traitCollectionDidChange`).

This seems very round-about when the Window already knows that the
resolution has changed.  However, this allows for handling the scale
change at various levels.  Another option would be to provide a window
hook when a window is created and listen in on the `WM_DPICHANGED` event
from the `Screen` or create an invisible Window to receive the message
even if the screen does not currently have a visible window.