Use `Foundation` as appropriate
Replace `Range` with `NSRange`.  As `Range` as designed to be similar to
`NSRange`, this is merely a renaming and removal of `Range`.

Replace `NSClassFromName` with the official version from Foundation.
The version in `Foundation` is far more complete and better maintained
and safer than using the SPI from the standard library.