Support: introduce WindowText property wrapper
The window text property is very useful as it is used in a number of
places.  It is backed by an opaque storage that is handled by the
implementation.  Querying requires determining the size and then
allocating a buffer and retrieving a copy.

The `_Win32WindowText` property wrapper allows exposing the window
property itself through an alternate name.  This requires some extra
setup to pass along the view that the property wrapper is associated
with to allow setting up the `HWND`.  However, this allows for a single
implementation to be shared across all the various sites.