UI: fix a few window sizing issues
We would previously compute the frame of the top-level window
incorrectly: the window is created with `WS_VISIBLE` and so the
`GetWindowRect` would give the size of the non-client area and the drop
shadow.  Use `GetClientRect` and `ClientToScreen` to convert the window
frame to screen units.  Furthermore, this avoids the DPI virtualization
that is part of the `GetWindowRect`.

Reparent the subview earlier during the process.  This is important to
get the full state of the window correct.