UI: make `View.hWnd` implicitly unwrapped
We internally assume that the `hWnd` for the `View` is always valid.
The window should only be destroyed when the instance is being
deallocated.  As a result, the `hWnd` should always be non-nil.  This
previously caused issues as the `CreateWindowExW` would accidentally
return `nil` and we would not notice it the return type is marked as
nullable.