UI: improve View construction
Default to constructing the View offscreen.  When the window is made key
and visible, it will make the view visible by walking the child
hierarchy.  This prevents the windows from being temporarily visible
until parented.

Mark the windows as popup to avoid rendering a non-client area and then
remove the thick frame style.  This properly renders the control when
built un-parented and then parented to the destination view.

Switch the `TextView` to a rich edit control making it more flexible and
renders the border better as well.