UI: sink `font` property to `View`
Make the `font` property a property of `View` rather than each control
implementing it.  Controls which expose control over the font can then
override the property and forward the `View` property which is limited
to `internal` access.

Furthermore simplify the property by making it a normal stored property
with a `didSet` observer.  This will become more useful as a subsequent
change will initialise the property always to ensure that the defaulted
font is the system font.