Calculator: adopt target action model for button handling
This migrates the Calculator demo to the action target model for
responding to the button presses.  This removes the vast majority of the
Win32 specific knowledge from the demo application.  The last piece that
remains at this point is the `CW_USEDEFAULT` used to place the window
where the window manager believes that the Window should be placed.