UI: add support for placeholder text in `TextField`
Since we model the `TextField` using a RichEdit, we cannot use the
simpler approach of using `EM_SETCUEBANNER` to render the placeholder
text.  Instead, rely on the fact that we subclass the window to override
the `WM_PAINT` event and render the placeholder text manually on top of
the default rendering.