UI: remove `WindowDelegate`
This was no longer actually used.  With the new responder chain in
progress, it will be possible to dispatch the even through the window
hierarchy and let the proper handler be triggered.  More importantly,
the previous target action model has already enabled handling button
primary actions.  This was purely vestigial code now.