UI: namespace `window` further into a variable (NFCI)
This adds a variable to encapsulate most of the Win32 specific window
details (window style and window class).  Ideally, the `hWnd` for the
View would also be encapsulated into the `win32` variable freeing up any
overlap between the rest of the API and making it easier to explicitly
delineate the Win32 specifics allowing the possibility of trying to
re-use some of the application life cycle (and view controller handling
in the future) with a different backend (e.g. GTK+).