UI: reorder the view hierarchy management declarations (NFCI)
This gathers the view hierarchy related functions and variables.  It
also adds a new variable which is the `window` to which the view
belongs.  Note that this is currently not available for use.  These are
all read-only externally and are only writable within the class, so
remove the misleading comment.