Application: correct the order of callbacks
This corrects the startup events to make the scene setup later and
correct the extension points to execute at the right time to enable
applications to configure the scene.  It temporarily removes the
tear down path as it still does not seem to entirely make sense.  The
scene and scene delegate is now configurable via code.