UI: `View` conforms to `TraitEnvironment`, `ContentSizeCateogryAdjusting`
`View` should conform to `TraitEnvironment` since it is impacted by the
current device traits.  This also allows `Label`, `TextField`, and
`TextView` implement the `ContentSizeCategoryAdjusting` conformance by
being able to observe trait environment changes, enabling them to adjust
the font if necessary.  This requires the addition of `FontMetrics` to
perform the scaling.
`FontMetrics.scaledFont(for:maximumPointSize:compatibleWith:)` needs to
be implemented still and is left for future work.