UI: handle `WM_GETMINMAXINFO` in `Window`
By intercepting this in the subclass procedure for Window, we can
restrict the client area sizing.  This allows us to honour the window
size request from the application as specified by the scene delegate.

Note that the client area must be adjusted to the window area when
limiting the tracking size of the Window as otherwise, the non-client
area will eat into the window shrinking the available client area.