UI: prevent resizing of the window in fixed mode
If the scene restricts the window to the same minimum and maximum
dimensions, prevent the user from resizing the window bounds.  We would
previously remove the minimize and maximize buttons but would still
silently allow the user to resize the window bounds from the edge.
Unset the `WS_SIZEBOX` window style to prevent this.