Create windows.yml
Add an initial GitHub Action to build with Swift Package Manager on PR requests.  This paves the way to promoting support for Swift Package Manager from experimental to available.  Although CMake is still recommended as the build system as it enables copying additional files which is not yet possible via Swift Package Manager.