UI: refactor, simplify menu handling
Split out the separator handling for `Win32MenuElement` into an
extension, reducing the amount of content in the main definition.

Replace the `of` "constructor" to a `convenience init` on
`Win32MenuElement`.

Rename the `win32ContextMenu` to `menu` in `View`.

Ad a generic helper to identify and cast the interaction from the view's
interactions.

Simplify the `WM_CONTEXTMENU` routine by always clearing the `menu`
field, replacing the menu, avoiding two separate paths.  Collapse the
access of the actions from the context menu configuration provided by
the delegate into a single chained query.  If one is provided, then only
do we create a new menu.