UI: avoid circular references
The parent view has a strong reference to the child view.  However, the
child view also has a strong reference to the parent.  This constructs a
circular reference.  Make the parent reference weak as the parent window
owns the child.  We similarly want to avoid the strong reference to the
Window object as the Window object owns the reference to the views.