build: update the Swift install rule
The Swift support in CMake has improved to the point where the only
missing piece is the support for `install(TARGET)`.  Emulate the
behaivour with `_install_target`.  Eventually, hopefully, this support
can be replaced with `install(TARGET)`.  Use the 5.4 support for
installing the module in the same manner on all targets.