Autolayout: implement `active` handler
When a constraint is activated, the constraint shall be added to the
nearest common ancestor.  Inactivation of the constraint does the same,
except instead of adding the constraint, removes it.

This implementation uses a poor scan for finding the nearest ancestor.
However, this should be possible to improve in the future.