tests: move LinuxMain.swift to allow exclusion
This attempts to adjust the layout of the tree as we do not want to use
the explicit manifest, instead preferring the test auto-discovery when
building with SPM.  The explicit manifest is only required for the CMake
based build.  The main.swift is entirely ignored by SPM as it is no
longer named `LinuxMain.swift`.