WinSDK: add an overload for `WS_CAPTION`
`WS_CAPTION` falls into the same category as the other window style
values - it is meant to be a DWORD.  In C/C++, the type is implicitly
converted, however, in Swift, it will not be.  Provide a DWORD overload
of the value.  For the cases where we need the value as a `LONG`, we can
explicitly use the `WinSDK.` prefix to retrieve the long typed value.
This cleans up most of the usage of the value.