PickerView: add an initial implementation of PickerView
This adds a constrained implementation of PickerView.  It sketches out
the various pieces that are needed to get a working picker view.
Ideally, we would use the `CBS_OWNERDRAWVARIABLE` to render the views
individually sized, but that seems to cause an issue with the draw
operation, where the draw is never invoked for each item.  This
currently restricts us to equal sized views rather than variable sized
views.

The other important missing piece here is the lack of support for
multi-component pickers.  This should be possible to do in practice
since the backing widget tree is a ListBox and an Edit container.  We
should be able to render more complicated views accordingly.

Resolves: #425