TableView: make TableViewCells visible when rendering them
This fixes the appearance of TableViewCells which are dynamically realized.  It raises additional questions for what the correct visibility of Views is: should reparenting implicitly make the view visible?  Should all views be implicitly be visible except for the Window?  That is out of scope for this particular change, which simply corrects the visibility of TableViewCell within the TableView.

Authored-by: Ryan Allan <ryan.allan@mts.com>