Label: partially implement `sizeThatFits(_:)`
Add a partial implementation for `sizeThatFits(_:)` for `Label`.  This
computes the extent of the rendered text in the currently specified
font.  This is missing support for frame margins and padding which
should be used to actually increase the size that fits, however, those
properties are not yet available on the type, making it not possible to
accommodate them in the current state.

The biggest missing piece here is support for multi-line titles and
support for attributed strings which are currently unsupported in the
implementation as well.