Views and Controls: mark Label and ImageView as not having interactions
The default state of these objects is meant to have user interactions
disabled.  This updates the implementation to mark them appropriately.
This will break the current behaviour of the Label which will accept
input.  However, the longer term desire is for tap events to be detected
using TapGestureRecognizer, which when bound, requires the user enable
the user interactions.  This moves us in that direction.