Update README.md
Update SPM instructions to avoid copying the manifest.  Instead, have users actually use the manifest tool to embed the manifest as a resource in the executable.  This currently requires the Windows manifest tool as the LLVM manifest tool does not support the `-outputresource:` option which is required to embed the new resource.