build: use glob for dependent modules
The dependent modules are meant to be pristine and do not impact the
build.  When building with CMake, we do not build the dependencies via
SPM.  Because the dependent projects have removed their CMake based
builds, we just simply re-create the sub-project here.  Because there
are meant to be no changes to these dependencies when building
Swift/Win32, we can simply glob the sources.  This avoids maintaining an
out-of-tree build for the project and manually updating the source list.
Furthermore, this matches the SPM based build more accurately.