View: add the `bounds` property
The bounds represents the coordinate space of the view.  It is the
region of the view which is actually rendered.  We do not currently
support setting the bounds, which allows for resizing the view around
the center by setting the size.  The origin of the view bounds may also
be altered, which will result in the area outside of the bounds being
clipped.