View: add `convert(_:to:)` for `Point`
This adds one of the view conversion routines.  The vast majority of the
change is the exposition on the mapping of the points as there is some
trickery involved here.  The operation implementation is split into
three cases:

- the point is in the correct coordinate space already:
    nothing to do
- the point is in either the immediate parent or child coordinate space:
    fast path the conversion since it is a simple change of basis
- the point is in some view hierarchy:
    convert the locations of the two views in the window based coordinate
    system and compute the transformation from that position.

The current implementation supports a single point, but the
implementation of the other three should be trivial as they are just the
same function called for different points or parameter order.