Examples: remove `HelloWorld`
Remove the `HelloWorld` project.  This was always meant to be closer to
a project template.  This is now deprecated and best handled in an
external repository at
https://github.com/compnerd/swift-win32-application.

The template project provides the appropriate structure without the
additional complexity of a general purpose project.  It should allow for
an easier time for someone to get started.