Control: functional handling for event triggering
Filter the triggering events first then invoke the handlers.  This is
slightly slower than doing it interspersed and may take a little more
memory (which can be handled by using `lazy`), but makes it easier to
reason about.  It hoists the event list out as Action bindings result
in a third use.