workflows: mock up a code-coverage workflow (#562)
This sets up the initial workflow for getting code-coverage information
from the build.  Because we want to track the changes overtime, the
uploaded status should be based on the merged changes, though it would
be nice to get indication on how a change influences the coverage.

This is currently disabled, requiring manual invocation; it also adds a few
workarounds until an updated snapshot can be rolled out.