Touches, Presses, and Gestures: flesh out `Press`
This adds the missing interfaces in `Press`, extracts the subtypes into
extensions, and adds a conformance for equatable and hashable.  The
equality if currently based on pointer identity, though it is
questionable if the equality should be value based.  The hashable
conformance simply adds the members.

`Press` is also now `open` to allow derivations by `open` types.