Menus and Shortcuts: add `Equatable` and `Hashable` conformances
Make `Menu` conform to `Equatable` and `Hashable`.  This is required for
`Menu` to be used in an `OrderedSet`.  Technically, the type can inherit
from `NSObject`, which gives it an implicit `Equatable` and `Hashable`
conformance.  Since we do not pervasively use `NSObject`, we must
synthesize the conformance explicitly.

Tweak the comments for `Press`'s conformances so that we can find all of
them with a single pattern.