Platform: start implementing a `MenuBuilder` for Win32
This is meant to be the internal implementation for a Windows menu
system/builder interface.  When complete, this should be able to subsume
the current implementation of `Win32Menu` and `Win32MenuElement`.  This
implementation instead keeps everything in terms of the raw `Menu` type
for the duration of the builder's lifetime.  When the builder is
complete, we can take the description as provided by the fluent
MenuBuilder API and materialize the menu structure in a single pass.
This will avoid the unnecessary allocation of handles during the
construction.