Responder: improve the responder hierarchy configuration
This improves the responder hierarchy handling to be more correct.  One
of the largest changes occurs in the `ViewController` type which gains
more of its interfaces.  These are needed for presenting a view
controller and to have the parent-child relationship for the view
controllers.  These impact the view hierarchy and thus must be present.