Drawing: add new constructor for `Color`
Add the `Color(dynamicProvider:)` constructor for `Color` and update the
various instances to use the new constructor to provide the dynamically
adjusting color.  The conversion to `COLORREF` will take the current
trait, which preserves the status quo behaviour, but enables future
enhancements to dynamically adjust to the traits for the rendering view.
Take the opportunity to add some additional documentation for the
various static members.