View and Controls: remove context menu handling
This was not completed, and more importantly, took over a name that is
meant to be used in subclasses.  This renames the `menu` item to
`hMenu`, and makes it a menu handle.  The construction of the menu needs
to be plumbed through, but this removes the old handling while the new
handling is implemented.