tests: ensure that the setup succeeded when testing
We need to ensure that the richedit module is loaded before we attempt
to use the richedit based controls.  This is normally taken care of
during the application setup, but because the tests do not run as an
application, we need to handle this in the test harness setup.  The
tests should subsequently verify that the module was loaded.