Platform: avoid warning from use of `_strerror_s`
The buffer size for `_strerror` and `_wcserror` are fixed at 94
characters.  Use this to allocate the fixed size buffer and build the
error message using `_wcserror_s` to avoid the warning as the allocation
should be small enough to get stack promoted.

Add a workaround for the 5.6.3 release where the Debug Info Verifier
will abort due to invalid debug info generation by bumping the minimum
to 5.7.