Views and Controls: support background colours
Generate a solid brush when we assign a background colour to a view.
This will then be used to erase the background to a solid fill when
rendering the view. This is sufficient in some cases but not in others.
The application of the null brush should give us a transparent view but
seems to not function properly.