Images and PDF: explicitly use WIC2
Explicitly specify the version of the Windows Imaging Component to avoid
an ambiguous reference. Swift recently introduced the ability to import
aliasing macros. This now creates a second wrapper declaration that
overloads the unversioned alias. Windows 10 implicitly uses the new
version, so we can expand that macro manually and avoid the ambiguous
reference.